requirements.txt
python-binance==0.4.1
krakenex==2.0.0
plotly==2.2.2
